Fix performance regression caused by xassert() macro
Before this commit, the xassert() macro still expanded to e.g.:

    if (unlikely(!(cond)))
        __builtin_unreachable();

...even when NDEBUG was set.

If `cond` happened to contain a function call that the compiler couldn't
prove to be side-effect free, this call would still have to be emitted
in the generated code, even when everything else could be optimized out.

This commit changes xassert() to be more in-keeping with the standard
assert() macro when NDEBUG is set; specifically, it causes the entire
statement to expand to nothing, which thus guarantees that any function
calls in the condition will be dropped, regardless of whether or not
they have side-effects.