term: cache currently selected charset
vt.c::action_print(), which is in the hot path, needs to lookup the
currently selected charset.

This currently requires two loads; first to load the current
charset *index*, and then another to get _which_ charset is currently
mapped to that index.

By caching the current charset, we only need one load.

We need to update the cache when:

* A different charset index is selected (i.e. SO/SI)
* The charset for the current index is changed (e.g. ‘\E(0’, ‘\E(B’
  etc)
* We restore saved charsets