Revert "vt: make anywhere() function return `term->vt.state` by default"
The loop in vt_from_slave() doesn't update `term->vt.state` on each
iteration and therefore may be incorrect when accessed by the anywhere()
function. Since this function is the only place other than the outer
loop where this field is read and since updating it per-iteration seems
to have quite some overhead, we instead just revert back to passing the
default return value to anywhere(), as was done prior to e4ff8d83d16.

This reverts commit e4ff8d83d1628c2fc92b14c7b0845215962fcbfc.