term: wait for client to die *before* tearing down the window
Waiting for the client application to terminate can take a noticeable
amount of time.

If we do this in a foot --server instance, we’ll be blocking other
footclient instances from processing any kind of events, including
Wayland events.

As an example, on a tiling WM with two footclient windows next to each
other, closing one of them _may_ cause the other window to be
resized (by Sway), before foot has been able to provide a properly
resized surface, resulting in the old (non-resized) surface being
centered in a too-large window by Sway.

This patch moves the whole client termination business from
term_destroy(), to term_shutdown(). This means it now runs *before*
the Wayland window is unmapped. This should allow other footclients to
respond in time once the window *is* unmapped.