osc8: handle LCF correctly
When the client application has printed something to the last column,
the cursor remains in that last column, but the LCF flag is set. The
next time something is printed, we linewrap.

This needs special consideration both when an OSC-8 URL is started,
and closed, when LCF is set.

Before this patch, emitting an OSC-8 URL while the cursor was in the
last column (with LCF set) would incorrectly include the last
character from that line, when the URL really should start at the next
line.

Furthermore, ending an OSC-8 URL with LCF set would inorrectly exclude
the last character.

This patch fixes this by using the will-be-printed-to
coordinates (i.e. the next line) in the OSC-8 start/end coordinates.

Note that this means that we _can_ point to uninitialized grid
data (i.e. to a NULL-row). This should be ok though, since either
something *will* be printed to that row before the URL is closed, or
a cursor movement causes the URL to be closed while still being empty,
which is a no-op.