selection: fix inconsistency in block selections
Previously, if you made a block selection, creating a single column
would set the left/right direction of a selection. This meant that:

1. Creating a single column and then dragging right created a selection
   which excluded the original column

2. Dragging your selection left and right would create a selection two
   columns wide for two consecutive cells of cursor movement

So, leave the selection as SELECTION_UNDIR when appropriate and only set
a pivot when the start and end columns are different. This makes the
behaviour the same as char-wise selections on a single line.