input: ignore “keyboard key” events on seats without keyboard focus
There are two ways this can happen:

1. The compositor did not send a “keyboard enter” event before sending
   the key release event.
2. The compositor did send a “keyboard enter” event, but did so before
   sending a “keyboard map” event. In this case, the seat’s XKB context
   hasn’t yet been set, and foot ignores the “keyboard enter” event.

Regardless of the root cause, this causes foot to crash because the
seat instance in foot isn’t able to track which terminal window
instance is being focused.

While this is a compositor bug, it still makes sense to detect this in
foot, and ignore such events, to ensure the user doesn’t lose any
work.

Closes #1097