vt: improve handling of codepoint 0xfe0f when grapheme-width-method != double-width
The Unicode codepoint 0xfe0f (Variation Selector 16) means the
preceding emoji should be rendered with its graphical representation,
rather than its text representation.

This typically means a double-width glyph.

Before this patch, we only handled this correctly for
grapheme-width-method=double-width.

In all other cases, we treated it as any other codepoint. Meaning, for
‘max’, we called wcwidth(0xfe0f), and compared against the last width.

For ‘wcswidth’ (the default), we also called wcwidth(0xfe0f) and added
this to the existing grapheme width.

With this patch, ‘max’ is changed to treat 0xfe0f as if it has width
== 2. So if the preceding emoji has width=1 (typically so), then we
now assign it width 2.

For `wcswidth`, we simply ignore the existing grapheme width and
simply set it to two. This will be correct when 0xfe0f is used
correctly, but may yield strange results otherwise.