sixel: limit image size to the one specified in the raster attrs
If a sixel has set an explicit width/height, in its raster attributes,
do *not* allow the image to extend beyond this.

Before this patch, we dynamically resized images if the sixel encoder
emitted sixel data beyond the previously set width/height.

Now, we ignore writes outside the configured size.